# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Rest API documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/passer": {
            "get": {
                "tags": [
                    "passer"
                ],
                "summary": "Get all passers for REST API",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "passerindex",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "available",
                                "pending",
                                "sold"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/passer/create": {
            "post": {
                "tags": [
                    "passer"
                ],
                "summary": "create all passers for REST API",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "passerstore",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "available",
                                "pending",
                                "sold"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "description": "Book data that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_examen": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "codeApogee": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "id_local": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "num_exam": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "isPresent": {
                                        "type": "boolean",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/passer/edit/{id_examen}/{id_local}/{codeApogee}": {
            "put": {
                "tags": [
                    "passer"
                ],
                "summary": "update all passers for REST API",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "passerupdate",
                "parameters": [
                    {
                        "name": "id_examen",
                        "in": "path",
                        "description": "examen id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id_local",
                        "in": "path",
                        "description": "local id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "codeApogee",
                        "in": "path",
                        "description": "codeApogee",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Book data that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_examen": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "codeApogee": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "id_local": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "num_exam": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "isPresent": {
                                        "type": "boolean",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/passer/{passer}": {
            "delete": {
                "tags": [
                    "passer"
                ],
                "summary": "delete all passers for REST API",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "passerdelete",
                "parameters": [
                    {
                        "name": "passer",
                        "in": "path",
                        "description": "passer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Input your Bearer token in the following format - Bearer {your_token_here}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": {
        "securityDefinitions": {
            "Bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        },
        "security": [
            {
                "Bearer": []
            }
        ]
    }
}